Fix for "Bad file descriptor" with --connect-timeout option (3.0)
This fixes #1834 and #1435 for branch 3.0. (3.1dev: #2047)

The --connect-timeout option had the problem that under certain circumstances
like parallel mass scanning didn't work. The culprit was that a subshell command
was used to connect to the target but the file descriptor wasn't exported.

The PR changes the logic so that the connect pre-check is still done in a subshell
If this fails it proceeds with error handling if NR_SOCKET_FAIL is above threshold.
Otherwise it just connects again.