Add --abort-no-tls switch
When used in automated pipelines, processing output from other tools
with testssl might cause a halt of the whole pipeline if the previous
output is detected false positive as TLS/SSL enabled Server.

Enabling this switch will cause testssl to return ALLOK after the first
sign, that TLS/SSL is not enabled on the server.