Fix error when no *.pem CA bundles exist
When there are no CA bundles, the lack of nullglob causes "$CA_BUNDLES_PATH/*.pem" to expand to itself, so that `basename` is called with "$CA_BUNDLES_PATH/*" and fails due to too many operands.

Compounded with the fact that *.pem files are in .gitignore, it's too easy to accidentally lose all CA bundles and trigger this error.